+++ packages/ubs/public/ReportViewer.js
@@ -33,10 +33,11 @@ td,th {
     a {
       text-decoration: none
     }
+    @page { margin: 1cm; }
+    #Header, #Footer { display: none !important; }
 }`
 const tableResizeCSS = '@media screen{th { resize: both; overflow: auto; }}'
-const printRepCSS = '@page{margin: 0mm;}' // this affects the margin in the printer settings
-const repCSS = baseRepCSS + tableResizeCSS + printRepCSS
+const repCSS = baseRepCSS + tableResizeCSS
 const HTML_PAGEBREAK_RE = new RegExp('<!-- pagebreak -->', 'gi')
 const HTML_PAGEBREAK_EL = '<img src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7" class="mce-pagebreak"/>'